Use EndpointMetadata to check for existing MapToApiAttribute at runtime
